requirements.txt
pandas==2.0.3
numpy==1.24.3
streamlit==1.28.0
plotly==5.17.0
krakenex==2.1.0
python-dotenv==1.0.0
requests==2.31.0
textblob==0.17.1
tqdm==4.66.1
ta==0.10.2
scikit-learn==1.3.0
joblib==1.3.2
python-binance==1.0.17
openai==0.28.0